@@ -1,6 +1,6 @@
 {
   "name": "squac-ui",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --host 0.0.0.0",
@@ -13,56 +13,59 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.1.5",
-    "@angular/cdk": "^10.2.4",
-    "@angular/common": "^10.1.5",
-    "@angular/compiler": "^10.1.5",
-    "@angular/core": "^10.1.5",
-    "@angular/forms": "^10.1.5",
-    "@angular/material": "^10.2.4",
-    "@angular/platform-browser": "^10.1.5",
-    "@angular/platform-browser-dynamic": "^10.1.5",
-    "@angular/router": "^10.1.5",
-    "@asymmetrik/ngx-leaflet": "^6.0.1",
-    "@asymmetrik/ngx-leaflet-draw": "^5.0.1",
-    "@casl/ability": "^4.1.0",
-    "@casl/angular": "^4.1.1",
-    "@swimlane/ngx-charts": "^14.0.0",
-    "@swimlane/ngx-datatable": "^17.0.0",
-    "@types/d3": "^5.7.2",
-    "angular-gridster2": "^9.1.0",
-    "core-js": "^2.6.11",
-    "d3": "^5.16.0",
+    "@angular/animations": "^13.1.0",
+    "@angular/cdk": "^13.1.0",
+    "@angular/common": "^13.1.0",
+    "@angular/compiler": "^13.1.0",
+    "@angular/core": "^13.1.0",
+    "@angular/forms": "^13.1.0",
+    "@angular/material": "^13.1.0",
+    "@angular/platform-browser": "^13.1.0",
+    "@angular/platform-browser-dynamic": "^13.1.0",
+    "@angular/router": "^13.1.0",
+    "@asymmetrik/ngx-leaflet": "^8.1.0",
+    "@asymmetrik/ngx-leaflet-draw": "^7.0.0",
+    "@casl/ability": "^5.4.3",
+    "@casl/angular": "^6.0.0",
+    "@swimlane/ngx-charts": "^20.1.0",
+    "@swimlane/ngx-datatable": "^20.0.0",
+    "@types/d3": "^7.1.0",
+    "@types/jasmine": "^3.10.2",
+    "angular-gridster2": "^13.1.0",
+    "d3": "^7.2.1",
+    "dayjs": "^1.10.7",
     "leaflet": "^1.6.0",
     "leaflet-draw": "^1.0.4",
-    "moment": "^2.27.0",
-    "ngx-daterangepicker-material": "^3.0.1",
-    "rxjs": "^6.5.5",
-    "serialize-javascript": "^3.1.0",
-    "timelines-chart": "^2.10.0",
+    "ng-mocks": "^13.3.0",
+    "ngx-daterangepicker-material": "^5.0.1",
+    "rxjs": "^7.4.0",
+    "timelines-chart": "^2.11.2",
+    "ts-node": "^10.4.0",
     "tslib": "^2.0.0",
-    "zone.js": "^0.10.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1001.6",
-    "@angular/cli": "^10.1.6",
-    "@angular/compiler-cli": "^10.1.5",
-    "@angular/language-service": "^10.1.5",
-    "@types/jasmine": "^2.8.17",
-    "@types/jasminewd2": "^2.0.8",
-    "@types/leaflet": "^1.5.19",
-    "@types/leaflet-draw": "^1.0.3",
-    "@types/node": "^12.12.67",
-    "codelyzer": "^5.2.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "@angular-devkit/build-angular": "^13.1.1",
+    "@angular-eslint/builder": "13.0.1",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular-eslint/eslint-plugin-template": "13.0.1",
+    "@angular-eslint/schematics": "13.0.1",
+    "@angular-eslint/template-parser": "13.0.1",
+    "@angular/cli": "^13.1.1",
+    "@angular/compiler-cli": "^13.1.0",
+    "@angular/language-service": "^13.1.0",
+    "@types/jasminewd2": "^2.0.10",
+    "@types/leaflet": "^1.7.6",
+    "@types/leaflet-draw": "^1.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.3.0",
+    "@typescript-eslint/parser": "^5.3.0",
+    "eslint": "^8.2.0",
+    "jasmine-spec-reporter": "^7.0.0",
+    "karma": "~6.3.9",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.3"
+    "typescript": "^4.5.3"
   }
 }
