@@ -1,8 +1,8 @@
-import { validateHorizontalPosition } from '@angular/cdk/overlay';
+import { STEPPER_GLOBAL_OPTIONS } from '@angular/cdk/stepper';
 import { Component, Inject, OnDestroy, OnInit } from '@angular/core';
-import { FormArray, FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { FormArray, FormBuilder, FormGroup, Validator, Validators } from '@angular/forms';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { ActivatedRoute, ActivationEnd, Params } from '@angular/router';
+import { MAT_FORM_FIELD_DEFAULT_OPTIONS } from '@angular/material/form-field';
 import { ChannelGroup } from '@core/models/channel-group';
 import { Metric } from '@core/models/metric';
 import { MessageService } from '@core/services/message.service';
@@ -10,15 +10,21 @@ import { Monitor } from '@features/monitors/models/monitor';
 import { Trigger } from '@features/monitors/models/trigger';
 import { MonitorsService } from '@features/monitors/services/monitors.service';
 import { TriggersService } from '@features/monitors/services/triggers.service';
-import { UserService } from '@features/user/services/user.service';
 import { merge, Subscription } from 'rxjs';
-import { switchMap, tap } from 'rxjs/operators';
+import { switchMap } from 'rxjs/operators';
 import { MonitorEditEntryComponent } from '../monitor-edit-entry/monitor-edit-entry.component';
 
 @Component({
   selector: 'app-monitor-edit',
   templateUrl: './monitor-edit.component.html',
-  styleUrls: ['./monitor-edit.component.scss']
+  styleUrls: ['./monitor-edit.component.scss'],
+  providers: [
+    // {provide: MAT_FORM_FIELD_DEFAULT_OPTIONS, useValue: {hideRequiredMarker: true}},
+    {
+      provide: STEPPER_GLOBAL_OPTIONS,
+      useValue: {displayDefaultIndicatorType: false},
+    }
+  ]
 })
 export class MonitorEditComponent implements OnInit, OnDestroy {
   subscriptions: Subscription = new Subscription();
@@ -26,9 +32,12 @@ export class MonitorEditComponent implements OnInit, OnDestroy {
   editMode: boolean;
   orgId: number;
   monitor: Monitor;
-  monitorForm: FormGroup;
   channelGroups: ChannelGroup[];
   metrics: Metric[];
+  emailValidator: Validators = Validators.pattern(/^([\w+-.%]+@[\w-.]+\.[A-Za-z]{2,4},?)+$/)
+  hideRequiredMarker = true;
+  floatLabel='always';
+  removeTriggerIDs=[];
   constructor(
     private formBuilder: FormBuilder,
     private monitorsService: MonitorsService,
@@ -39,67 +48,120 @@ export class MonitorEditComponent implements OnInit, OnDestroy {
   ) { }
 
   intervalTypes: string[] = ['minute', 'hour', 'day'];
-  stats: string[] = ['count', 'sum', 'avg', 'min', 'max'];
-  levels: number[] = [1, 2, 3];
+  stats: object[] = [
+    { value: 'count', name: 'count'},
+    { value : 'sum', name: 'sum'},
+    { value : 'avg', name: 'average'},
+    { value : 'min', name: 'minimum'},
+    { value : 'max', name: 'maximum'}];
+  value_operators: object[] = [
+    { value: 'outsideof', name: 'outside of'},
+    { value: 'within', name: 'within'},
+    { value: '==', name: 'equal to'},
+    { value: '<', name: 'less than'},
+    { value: '<=', name: 'less than or equal to'},
+    { value: '>', name: 'greater than'},
+    { value: '>=', name: 'greater than or equal to'}];
+  num_channels_operators: object[] = [
+    { value: 'any', name: 'any'},
+    { value: '==', name: 'exactly'},
+    { value: '<', name: 'more than'},
+    { value: '>', name: 'less than'}
+  ];
+
   // selectedType;
   // selectedStat;
   selectedChannelGroup: ChannelGroup;
   selectedMetric: Metric;
 
-  ngOnInit() {
-    this.monitorForm = this.formBuilder.group({
-      name: ['', Validators.required],
-      intervalCount: ['', [ Validators.required, Validators.min(1)]],
-      numberChannels: ['', [Validators.required, Validators.min(1)]],
-      intervalType: ['', Validators.required],
-      stat: ['', Validators.required],
-      metric: ['', Validators.required],
-      channelGroup: ['', Validators.required],
-      triggers: this.formBuilder.array([])
-    });
+  monitorForm = this.formBuilder.group({
+    name: ['', Validators.required],
+    intervalCount: ['', [ Validators.required, Validators.min(1)]],
+    intervalType: ['', Validators.required],
+    stat: ['', Validators.required],
+    metric: ['', Validators.required],
+    channelGroup: ['', Validators.required],
+    triggers: this.formBuilder.array([])
+  });
 
+  // Set up form fields
+  ngOnInit() {
+    let test = "test";
     this.metrics = this.data.metrics;
     this.channelGroups = this.data.channelGroups;
     this.editMode = !!this.data.monitor;
     this.initForm();
   }
-
+  
+  // Access triggers
   get triggers(){
     return this.monitorForm.get('triggers') as FormArray;
   }
 
-  addTrigger(trigger?: Trigger) {
-    this.triggers.push( this.formBuilder.group({
-      min: [trigger ? trigger.min : null],
-      max: [trigger ? trigger.max : null],
-      inclusive: [trigger ? trigger.bandInclusive : false],
-      level: [trigger ? trigger.level : null],
-      id: [trigger ? trigger.id : null]
-    }));
+  // Return first trigger, considered default
+  get defaultTrigger(){
+    return this.triggers.at(0).value;
   }
 
-  removeThreshold(index){
-    const trigger = this.triggers.at(index).value;
-    if (trigger.id) {
-      this.triggers.at(index).setValue({
-        id: trigger.id,
-        min: null,
-        max: null,
-        level: null,
-       inclusive: null
-      });
+  // Add trigger info to form
+  makeTriggerForm (trigger?: Trigger) {
+    return this.formBuilder.group({
+      val1: [trigger ? trigger.val1 : null, Validators.required],
+      val2: [trigger ? trigger.val2 : null],
+      id: [trigger ? trigger.id : null],
+      value_operator: [trigger ? trigger.value_operator : null, Validators.required],
+      num_channels:[trigger ? trigger.num_channels : null],
+      num_channels_operator:[trigger ? trigger.num_channels_operator : null, Validators.required],
+      alert_on_out_of_alarm: [trigger ? trigger.alert_on_out_of_alarm : false],
+      email_list:[trigger ? trigger.email_list : null, this.emailValidator]
+    });
+  }
+
+  // Add a new trigger
+  addTrigger(trigger?: Trigger) {
+    const triggerFormGroup = this.makeTriggerForm(trigger);
+    triggerFormGroup.valueChanges.subscribe(value => this.validateTrigger(value, triggerFormGroup))
+    this.triggers.push(triggerFormGroup);
+  }
+
+  validateTrigger(values, triggerFormGroup) {
+    const val2 = triggerFormGroup.get('val2');
+    const num_channels = triggerFormGroup.get('num_channels');
+    if(values.value_operator !== 'outsideof' && values.value_operator !== 'within') {
+      val2.setValue("", {emitEvent: false});
+      val2.disable({emitEvent: false});
+      val2.removeValidators(Validators.required, {emitEvent: false});
     } else {
-      this.triggers.removeAt(index);
+      val2.addValidators(Validators.required, {emitEvent: false});
+      val2.enable({emitEvent: false});
+    }
+    if(values.num_channels_operator === 'any') {
+      num_channels.setValue("", {emitEvent: false});
+      num_channels.disable({emitEvent: false});
+      num_channels.removeValidators(Validators.required, {emitEvent: false});
+    } else {
+      num_channels.enable({emitEvent: false});
+      num_channels.addValidators(Validators.required, {emitEvent: false});
     }
   }
+  // Remove given trigger
+  removeTrigger(index){
+    const trigger = this.triggers.at(index).value;
+    if (trigger.id) {
+      this.removeTriggerIDs.push(+trigger.id);
+    }
+    this.triggers.removeAt(index);
+  }
 
+  // Kill any open subs 
   ngOnDestroy(): void {
     this.subscriptions.unsubscribe();
   }
 
+  // Save monitor to SQUACapi
   save() {
     const values = this.monitorForm.value;
-
+    // console.log(values)
     const monitor = new Monitor(
       this.id,
       values.name,
@@ -107,7 +169,6 @@ export class MonitorEditComponent implements OnInit, OnDestroy {
       values.metric.id,
       values.intervalType,
       values.intervalCount,
-      values.numberChannels,
       values.stat,
       null,
       values.triggers
@@ -117,25 +178,28 @@ export class MonitorEditComponent implements OnInit, OnDestroy {
       monitor
     ).pipe(
       switchMap(m => {
-        return merge(...this.triggersService.updateTriggers(values.triggers, m.id));
+        return merge(...this.triggersService.updateTriggers(values.triggers, this.removeTriggerIDs, m.id));
       })
     ).subscribe(
-      success => {
-        this.messageService.message('Monitor saved.');
-        this.cancel();
-        // this.router.navigate()
-      },
-      error => {
-        this.messageService.error('Could not save monitor.');
+      {
+        next: () => {
+          this.messageService.message('Monitor saved.');
+          this.cancel();
+        },
+        error: () => {
+          this.messageService.error('Could not save monitor.');
+        }
       }
     );
   }
 
+  // Cancel and don't save changes
   cancel(monitor?: Monitor) {
     this.dialogRef.close(monitor);
     // route out of edit
   }
 
+  // Fill in metric info
   private initForm() {
 
     if (this.editMode) {
@@ -148,7 +212,6 @@ export class MonitorEditComponent implements OnInit, OnDestroy {
         {
           name: this.monitor.name,
           intervalCount: this.monitor.intervalCount,
-          numberChannels: this.monitor.numberChannels,
           intervalType: this.monitor.intervalType,
           stat: this.monitor.stat,
           channelGroup: this.selectedChannelGroup,
@@ -156,10 +219,15 @@ export class MonitorEditComponent implements OnInit, OnDestroy {
         }
       );
 
-      this.monitor.triggers.forEach((trigger) => {
+      this.monitor.triggers.forEach((trigger, i) => {
+        console.log(trigger)
         this.addTrigger(trigger);
       });
 
+    } else {
+      this.addTrigger(); //have default trigger ready
     }
+
+
   }
 }
