@@ -26,7 +26,6 @@ import { HomeComponent } from './core/components/home/home.component';
 import { BrowserModule } from '@angular/platform-browser';
 import { ConfigurationService } from '@core/services/configuration.service';
 
-
 export function initApp(configurationService: ConfigurationService) {
   return () => configurationService.load().toPromise();
 }
@@ -47,9 +46,8 @@ export function initApp(configurationService: ConfigurationService) {
     NoopAnimationsModule,
     BrowserAnimationsModule,
     BrowserModule,
-    LeafletModule.forRoot(),
-    LeafletDrawModule.forRoot(),
-
+    LeafletModule,
+    LeafletDrawModule,
     // Always load this at the end or the routing gets weird
     AppRoutingModule
   ],
