@@ -1,66 +1,49 @@
-import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { TestBed } from '@angular/core/testing';
 import { HeaderComponent } from './header.component';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { RouterTestingModule } from '@angular/router/testing';
-import { SquacApiService } from '../../services/squacapi.service';
-import { MockSquacApiService } from '../../services/squacapi.service.mock';
-import { MatToolbarModule } from '@angular/material/toolbar';
-import { Ability, PureAbility } from '@casl/ability';
-import { AbilityModule } from '@casl/angular';
-import { AppAbility } from '../../utils/ability';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { MessageService } from '@core/services/message.service';
 import { AuthService } from '@core/services/auth.service';
-import { MockAuthService } from '@core/services/auth.service.mock';
-import { UserService } from '@features/user/services/user.service';
-import { MockUserService } from '@features/user/services/user.service.mock';
-import { throwError } from 'rxjs';
+import { MockInstance, MockModule, MockProviders, MockRender, ngMocks } from 'ng-mocks';
+import { SharedModule } from '@shared/shared.module';
+import { AppAbility } from '@core/utils/ability';
+import { PureAbility } from '@casl/ability';
 
 describe('HeaderComponent', () => {
-  let component: HeaderComponent;
-  let fixture: ComponentFixture<HeaderComponent>;
-  let authService: AuthService;
-  let userService: UserService;
-  let messageService: MessageService;
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ HeaderComponent ],
+  ngMocks.faster();
+  MockInstance.scope();
+
+  beforeAll(async () => {
+    return TestBed.configureTestingModule({
       imports: [
-        HttpClientTestingModule,
-        RouterTestingModule,
-        MatToolbarModule,
-        AbilityModule,
-        MatSnackBarModule
+        MockModule(SharedModule)
+      ],
+      declarations: [
+        HeaderComponent
       ],
       providers: [
-        MessageService,
-        { provide: SquacApiService, useValue: MockSquacApiService },
-        { provide: AppAbility, useValue: new AppAbility() },
-        { provide: AuthService, useValue: new MockAuthService()},
-        { provide: PureAbility , useExisting: Ability },
-        { provide: UserService, useValue: new MockUserService()}
+        MockProviders(AuthService, AppAbility, PureAbility)
       ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(HeaderComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-    authService = TestBed.inject(AuthService);
-    userService = TestBed.inject(UserService);
-    messageService = TestBed.inject(MessageService);
+    }).compileComponents();
   });
 
   it('should create', () => {
-    expect(component).toBeTruthy();
+    const fixture = MockRender(HeaderComponent);
+    expect(
+      fixture.point.componentInstance,
+    ).toEqual(jasmine.any(HeaderComponent));
   });
 
   it('should logout', () => {
-    const authSpy = spyOn(authService, 'logout');
+    const spyLogout = MockInstance(
+      AuthService,
+      'logout',
+      jasmine.createSpy("logoutSpy")
+    );
+
+    const fixture = MockRender(HeaderComponent);
+
+    const component = fixture.point.componentInstance;
+
     component.logout();
-    expect(authSpy).toHaveBeenCalled();
+
+    expect(spyLogout).toHaveBeenCalled();
   });
 });
