@@ -1,34 +1,26 @@
-import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { TestBed } from '@angular/core/testing';
 import { HomeComponent } from './home.component';
 import { HeaderComponent } from '../header/header.component';
 import { RouterTestingModule } from '@angular/router/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { UserService } from '@features/user/services/user.service';
-import { MockUserService } from '@features/user/services/user.service.mock';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { MatToolbarModule } from '@angular/material/toolbar';
+import { MockComponent, MockInstance, MockModule, MockProvider, MockRender, ngMocks } from 'ng-mocks';
+import { MaterialModule } from '@shared/material.module';
+import { MessageService } from '@core/services/message.service';
 
 describe('HomeComponent', () => {
-  let component: HomeComponent;
-  let fixture: ComponentFixture<HomeComponent>;
+  ngMocks.faster();
+  MockInstance.scope();
 
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      imports: [RouterTestingModule, HttpClientTestingModule, MatSnackBarModule, MatToolbarModule],
-      declarations: [ HomeComponent , HeaderComponent],
-      providers: [{provide: UserService, useClass: MockUserService}]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(HomeComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
+  beforeAll(async () => {
+    return TestBed.configureTestingModule({
+      imports: [RouterTestingModule, MockModule(MaterialModule)],
+      declarations: [HomeComponent, MockComponent(HeaderComponent)],
+      providers: [MockProvider(MessageService)]
+    }).compileComponents();
   });
 
   it('should create', () => {
+    const fixture = MockRender(HomeComponent);
+    const component = fixture.point.componentInstance;
     expect(component).toBeTruthy();
   });
 });
