@@ -1,27 +1,23 @@
-import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
+import { MockInstance, MockRender, ngMocks } from 'ng-mocks';
 
 import { NotFoundComponent } from './not-found.component';
 
 describe('NotFoundComponent', () => {
-  let component: NotFoundComponent;
-  let fixture: ComponentFixture<NotFoundComponent>;
+  ngMocks.faster();
+  MockInstance.scope();
 
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ NotFoundComponent ],
+  beforeAll(async () => {
+    return TestBed.configureTestingModule({
+      declarations: [NotFoundComponent],
       imports: [RouterTestingModule.withRoutes([])]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(NotFoundComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
+    }).compileComponents();
   });
 
   it('should create', () => {
+    const fixture = MockRender(NotFoundComponent);
+    const component = fixture.point.componentInstance;
     expect(component).toBeTruthy();
   });
 });
