@@ -3,11 +3,13 @@ import { Dashboard } from '../../models/dashboard';
 import { ActivatedRoute, Router, Params } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { ViewService } from '@core/services/view.service';
-import * as moment from 'moment';
+import * as dayjs from 'dayjs';
+import * as utc from 'dayjs/plugin/utc';
+import * as timezone from 'dayjs/plugin/timezone';
 import { AppAbility } from '@core/utils/ability';
 import { DaterangepickerDirective } from 'ngx-daterangepicker-material';
 import { ConfirmDialogService } from '@core/services/confirm-dialog.service';
-import { ConfigurationService } from '@core/services/configuration.service';
+import { DateService } from '@core/services/date.service';
 
 //
 @Component({
@@ -20,12 +22,13 @@ export class DashboardDetailComponent implements OnInit, OnDestroy {
   dashboard: Dashboard;
   subscription: Subscription = new Subscription();
   status;
-  maxDate: moment.Moment;
+  startDate: dayjs.Dayjs;
+  maxDate: dayjs.Dayjs;
   error: string = null;
   unsaved = false;
   archiveType: string;
   archiveStat: string;
-  archiveStatTypes: string[] = ['min', 'max', 'mean', 'median', 'stdev', 'num_samps', 'p05', 'p10', 'p90', 'p95'];
+  archiveStatTypes: string[] = ['min', 'max', 'mean', 'median', 'stdev', 'num_samps', 'p05', 'p10', 'p90', 'p95', 'minabs', 'maxabs'];
   // TODO: make this a separate component, its making this too busy
   selected: {
     startDate,
@@ -33,7 +36,7 @@ export class DashboardDetailComponent implements OnInit, OnDestroy {
   };
   selectedRange: string;
   liveMode: boolean;
-  startDate: moment.Moment;
+
   // settings for date select
   locale;
   ranges = {};
@@ -49,20 +52,20 @@ export class DashboardDetailComponent implements OnInit, OnDestroy {
     private viewService: ViewService,
     private ability: AppAbility,
     private confirmDialog: ConfirmDialogService,
-    configService: ConfigurationService,
+    private dateService: DateService
   ) {
-    this.rangeLookUp = configService.getValue('dateRanges');
-    this.locale = configService.getValue('locale');
+    dayjs.extend(utc);
+    dayjs.extend(timezone);
+    dayjs.tz.setDefault('Etc/UTC');
   }
 
   ngOnInit() {
-
-    this.maxDate = moment.utc();
-    this.startDate = moment.utc();
+    this.rangeLookUp = this.dateService.dateRanges;
+    this.maxDate = this.dateService.now();
+    this.startDate = this.dateService.now();
     this.makeTimeRanges();
-
-
-    this.route.data.subscribe(
+    this.locale = this.dateService.locale;
+    const dashboardSub = this.route.data.subscribe(
       data => {
         this.status = 'loading';
         this.dashboard = data.dashboard;
@@ -77,17 +80,22 @@ export class DashboardDetailComponent implements OnInit, OnDestroy {
           this.archiveType = this.dashboard.archiveType;
           if (range && this.rangeLookUp) {
             this.selectedRange = this.rangeLookUp[range];
+            
             this.selected = {
-              startDate: this.ranges[this.selectedRange][0],
-              endDate: this.ranges[this.selectedRange][1]
+              startDate: this.dateService.subtractFromNow(range, 'seconds'),
+              endDate: this.dateService.now()
             };
+
           } else {
             this.selected = {
-              startDate: moment.utc(start),
-              endDate: moment.utc(end)
+              startDate: this.dateService.parseUtc(start),
+              endDate: this.dateService.parseUtc(end)
             };
             this.selectedRange = start + ' - ' + end;
           }
+
+          console.log(start, end)
+          console.log(this.selected)
           this.error = null;
         }
       }
@@ -109,8 +117,7 @@ export class DashboardDetailComponent implements OnInit, OnDestroy {
       }
     );
 
-    // this.subscription.add(dashSub);
-    // this.subscription.add(dashIdSub);
+    this.subscription.add(dashboardSub);
     this.subscription.add(statusSub);
     this.subscription.add(errorSub);
   }
@@ -118,7 +125,7 @@ export class DashboardDetailComponent implements OnInit, OnDestroy {
   makeTimeRanges(){
     for (const range in this.rangeLookUp) {
       if (this.rangeLookUp[range]) {
-        this.ranges[this.rangeLookUp[range]] = [moment.utc().subtract(+range, 'seconds'), this.startDate];
+        this.ranges[this.rangeLookUp[range]] = [this.dateService.subtractFromNow(+range, 'seconds'), this.startDate];
       }
     }
   }
@@ -130,25 +137,31 @@ export class DashboardDetailComponent implements OnInit, OnDestroy {
     this.viewService.setArchive(this.archiveType, this.archiveStat);
   }
   // FIXME: milliseconds of difference are causing it to not recognize
-  lookupRange(startDate: moment.Moment, endDate: moment.Moment): number | void {
-
+  lookupRange(startDate: dayjs.Dayjs, endDate: dayjs.Dayjs): number | void {
+    let diff = this.dateService.diff(endDate, startDate);
     // check if end of range close to now
-    if (Math.abs(endDate.diff(this.startDate)) < 1000 ) {
-
+    if (Math.abs(diff) < 1 ) {
       this.liveMode = true;
-      const diff = Math.round(endDate.diff(startDate) / 100000 ) * 100; // account for ms of weirdness
-      this.selectedRange = this.rangeLookUp[diff];
-      return diff;
+      const roundDiff = Math.round(diff / 100 ) * 100; // account for ms of weirdness
+      this.selectedRange = this.rangeLookUp[roundDiff];
+      return roundDiff;
     } else {
       this.liveMode = false;
-      this.selectedRange = startDate.format(this.locale.displayFormat) + ' - ' + endDate.format(this.locale.displayFormat);
+      this.selectedRange = this.dateService.displayFormat(startDate) + ' - ' + this.dateService.displayFormat(startDate);
     }
   }
 
-  datesSelected(chosenDate: {startDate: moment.Moment; endDate: moment.Moment }): void {
+  datesSelected(chosenDate: {startDate: dayjs.Dayjs; endDate: dayjs.Dayjs }): void {
+    let start = chosenDate.startDate;
+    let end = chosenDate.endDate
+
+    if(start && end) {
+      start = this.dateService.correctForLocal(start);
+      end = this.dateService.correctForLocal(end)
+    }
+
     this.unsaved = true;
-    const start = chosenDate.startDate;
-    const end = chosenDate.endDate;
+    console.log(start, end)
     if (start && end) {
       const range = this.lookupRange(start, end);
       this.viewService.datesChanged(
@@ -173,8 +186,9 @@ export class DashboardDetailComponent implements OnInit, OnDestroy {
   }
 
   // currently saves any time dates are changed, may want to move to a save button
-  selectDateRange(startDate: moment.Moment, endDate: moment.Moment, range?: number) {
-
+  selectDateRange(startDate: dayjs.Dayjs, endDate: dayjs.Dayjs, range?: number) {
+    this.selected.startDate = this.dateService.toUtc(startDate);
+    this.selected.endDate = this.dateService.toUtc(endDate);
   }
 
   deleteDashboard() {
