@@ -19,14 +19,13 @@ export class DashboardViewComponent implements OnInit, OnDestroy {
   constructor(
     private userService: UserService,
     private router: Router,
-    private route: ActivatedRoute,
-    private orgService: OrganizationsService
+    private route: ActivatedRoute
   ) {
 
   }
 
   ngOnInit() {
-    const dashboardsSub = this.route.params.subscribe(
+    const activeDashboardSub = this.route.params.subscribe(
       (params: Params) => {
         this.activeDashboardId = +params.dashboardId;
       },
@@ -35,9 +34,15 @@ export class DashboardViewComponent implements OnInit, OnDestroy {
       }
     );
 
-    if (this.route.snapshot && this.route.snapshot.data) {
-      this.dashboards = this.route.snapshot.data.dashboards;
-    }
+    const dashboardsSub = this.route.data.subscribe(
+      data => {
+        if (data.dashboards && data.dashboards.error) {
+          console.log('error in dashboard');
+        } else {
+          this.dashboards = data.dashboards;
+        }
+      }
+    );
 
     const userService = this.userService.user.subscribe(
       user => {
@@ -48,6 +53,7 @@ export class DashboardViewComponent implements OnInit, OnDestroy {
     // this.subscription.add(dashboardsService);
     this.subscription.add(userService);
     this.subscription.add(dashboardsSub);
+    this.subscription.add(activeDashboardSub);
   }
 
   ngOnDestroy(): void {
