@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
 import { Adapter } from '@core/models/adapter';
-import { Channel } from '@core/models/channel';
 import { ApiGetChannelGroup, ChannelGroup, ChannelGroupAdapter } from '@core/models/channel-group';
 import { Metric, ApiGetMetric, MetricAdapter } from '@core/models/metric';
 import { ApiGetTrigger, Trigger, TriggerAdapter } from './trigger';
@@ -13,7 +12,6 @@ export class Monitor {
     public metricId: number,
     public intervalType: string,
     public intervalCount: number,
-    public numberChannels: number,
     public stat: string,
     public owner: number,
     public triggers: Trigger[]
@@ -21,6 +19,10 @@ export class Monitor {
 
   channelGroup: ChannelGroup;
   metric: Metric;
+
+  static get modelName() {
+    return 'Monitor';
+  }
 }
 
 export interface ApiGetMonitor {
@@ -30,7 +32,6 @@ export interface ApiGetMonitor {
   metric: number | ApiGetMetric;
   interval_type: string;
   interval_count: number;
-  num_channels: number;
   stat: string;
   name: string;
   created_at: string;
@@ -44,7 +45,6 @@ export interface ApiPostMonitor {
   metric: number;
   interval_type: string;
   interval_count: number;
-  num_channels: number;
   stat: string;
   name: string;
 }
@@ -91,7 +91,6 @@ export class MonitorAdapter implements Adapter<Monitor> {
       metricId,
       item.interval_type,
       item.interval_count,
-      item.num_channels,
       item.stat,
       +item.user_id,
       triggers
@@ -107,7 +106,6 @@ export class MonitorAdapter implements Adapter<Monitor> {
     return {
       interval_type: item.intervalType,
       interval_count: item.intervalCount,
-      num_channels: item.numberChannels,
       channel_group: item.channelGroupId,
       metric: item.metricId,
       stat: item.stat,
