@@ -44,9 +44,7 @@ export class WidgetEditEntryComponent implements OnInit, OnDestroy {
           this.channelGroups = this.route.snapshot.data.channelGroups;
         }
 
-        this.widget = this.viewService.getWidget(this.widgetId);
-
-        if (this.widgetId && !this.widget){
+        if (this.widgetId){
           this.widgetsService.getWidget(this.widgetId).subscribe(
             widget => {
               this.widget = widget;
