@@ -2,19 +2,24 @@ import { waitForAsync, ComponentFixture, TestBed, inject } from '@angular/core/t
 
 import { WidgetEditComponent } from './widget-edit.component';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MaterialModule } from '@shared/material.module';
 import { MatDialogRef, MatDialog, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { NgxDatatableModule } from '@swimlane/ngx-datatable';
-import { ThresholdEditComponent } from './threshold-edit/threshold-edit.component';
-import { MetricsEditComponent } from './metrics-edit/metrics-edit.component';
-import { ChannelGroupsEditComponent } from './channel-groups-edit/channel-groups-edit.component';
-import { LoadingComponent } from '@shared/components/loading/loading.component';
 import { AbilityModule } from '@casl/angular';
 import { Ability, PureAbility } from '@casl/ability';
 import { AppAbility } from '@core/utils/ability';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { WidgetInfoEditComponent } from './widget-info-edit/widget-info-edit.component';
+import { WidgetEditService } from '@features/widgets/services/widget-edit.service';
+import { MockWidgetEditService } from '@features/widgets/services/widget-edit.service.mock';
+import { MessageService } from '@core/services/message.service';
+import { ChannelGroup } from '@core/models/channel-group';
+import { Metric } from '@core/models/metric';
+import { UserService } from '@features/user/services/user.service';
+import { MockUserService } from '@features/user/services/user.service.mock';
+import { OrganizationsService } from '@features/user/services/organizations.service';
+import { MockOrganizationsService } from '@features/user/services/organizations.service.mock';
+import { Component } from '@angular/core';
 
 describe('WidgetEditComponent', () => {
   let component: WidgetEditComponent;
@@ -29,6 +34,7 @@ describe('WidgetEditComponent', () => {
       imports: [
         HttpClientTestingModule,
         NoopAnimationsModule,
+        FormsModule,
         ReactiveFormsModule,
         MaterialModule,
         NgxDatatableModule,
@@ -40,20 +46,26 @@ describe('WidgetEditComponent', () => {
           useValue: mockDialogRef
         },
         { provide: MAT_DIALOG_DATA, useValue: {
-          data : {
-
-          }
+            channelGroups: [
+              new ChannelGroup(1, 1, 'name', 'Desc', 1, [1])
+            ],
+            metrics: [
+              new Metric(1, 1, 'name', 'code', 'Code', 'ref', 'unit', 1, 1, 0)
+            ]
         } },
         { provide: AppAbility, useValue: new AppAbility() },
-        { provide: PureAbility , useExisting: Ability }
+        { provide: PureAbility , useExisting: Ability },
+        { provide: WidgetEditService, useValue: new MockWidgetEditService()},
+        {
+          provide: UserService, useValue: new MockUserService()
+        },
+        {
+          provide: OrganizationsService, useValue: new MockOrganizationsService()
+        },
+        MessageService
       ],
       declarations: [
-        WidgetEditComponent,
-        ThresholdEditComponent,
-        MetricsEditComponent,
-        ChannelGroupsEditComponent,
-        LoadingComponent,
-        WidgetInfoEditComponent
+        WidgetEditComponent
       ]
     })
     .compileComponents();
@@ -66,7 +78,6 @@ describe('WidgetEditComponent', () => {
       fixture = TestBed.createComponent(WidgetEditComponent);
       component = fixture.componentInstance;
       fixture.detectChanges();
-
     }
     )
   );
