@@ -4,8 +4,9 @@
 
 export const environment = {
   production: false,
-  apiUrl: 'https://squacapi.pnsn.org/',
-  // apiUrl: 'https://staging-squacapi.pnsn.org/',
+  // apiUrl: 'http://localhost:8000/',
+  apiUrl: 'https://staging-squacapi.pnsn.org/',
+  // apiUrl: 'https://squacapi.pnsn.org/',
   version: 'v1.0/'
 };
 
@@ -16,4 +17,4 @@ export const environment = {
  * This import should be commented out in production mode because it will have a negative impact
  * on performance if an error is thrown.
  */
-// import 'zone.js/dist/zone-error';  // Included with Angular CLI.
+// import 'zone.js/plugins/zone-error';  // Included with Angular CLI.
